feat!: removes unused or deprecated types and utils
