+++ src/index.ts
@@ -1,11 +1,5 @@
 export { Passage } from './classes/Passage/Passage';
-export * from './types/AuthStrategy';
-export * from './types/JWKS';
-export * from './types/MagicLink';
-export * from './types/PassageConfig';
-export * from './types/User';
 export * from './generated/models';
-export * from './models';
 export * from './classes/Auth';
 export * from './classes/User';
 export * from './classes/PassageError';
