--- src/types/AuthStrategy.ts
@@ -1,8 +1,0 @@
-/* eslint-disable no-unused-vars */
-
-enum AuthStrategyEnum {
-    COOKIE = 'COOKIE',
-    HEADER = 'HEADER',
-}
-
-export type AuthStrategy = keyof typeof AuthStrategyEnum;
