--- src/types/JWKS.ts
@@ -1,10 +1,0 @@
-import { JWK } from 'jose';
-export interface AUTHCACHE {
-    [appID: string]: {
-        jwks: JWKS;
-    };
-}
-
-export interface JWKS {
-    [kid: string]: JWK;
-}
