--- src/types/MagicLink.ts
@@ -1,6 +1,0 @@
-/* eslint-disable */
-export enum ChannelEnum {
-    email = 'email',
-    phone = 'phone',
-}
-export type ChannelType = keyof typeof ChannelEnum;
